// frontend/src/types/CurrencyType.ts
export interface Currency {
	uuid: string
	symbol: string
	name: string
	color: string
	iconUrl: string
	marketCap: string
	price: string
	listedAt: number
	tier: number
	change: string
	rank: number
	sparkline: any
	lowVolume: boolean
	coinrankingUrl: string
	"24hVolume": string
	btcPrice: string
}